Finished implementation of euler and midpoint integrators
added functionality for switching between integrators
added physics function library in the "Physics" namespace
switched computation to double precision

+++ Simulations/EulerParticleIntegrator.h
@@ -8,6 +8,6 @@ class EulerParticleIntegrator :
 	public ParticleIntegratorBase
 {
 
-	virtual WorldState GetNextSimulationStep(const WorldState& CurrentWorldState, float deltaSeconds) override;
+	virtual WorldState GetNextSimulationStep(const WorldState& CurrentWorldState, double deltaSeconds) override;
 };
