+++ Simulations/IParticleIntegrator.h
@@ -13,7 +13,7 @@ public:
 	 * @param deltaSeconds the time which should pass in the world
 	 * @return the new world state after deltaSeconds time has passed in the world
 	 */
-	virtual WorldState GetNextSimulationStep(const WorldState& CurrentWorldState, float deltaSeconds) = 0;
+	virtual WorldState GetNextSimulationStep(const WorldState& CurrentWorldState, double deltaSeconds) = 0;
 
 	/**
 	 * adds a vector to the global forces which will be applied to every particle during the simulation
